+++ client/app/pods/paper/workflow/discussions/new/route.js
@@ -4,14 +4,5 @@ import AuthorizedRoute from 'tahi/routes/authorized';
 
 export default AuthorizedRoute.extend(DiscussionsNewRouteMixin, {
   // required to generate route paths:
-  subRouteName: 'workflow',
-  can: Ember.inject.service('can'),
-
-  beforeModel(transition){
-    this.get('can').can('start_discussion', this.modelFor('paper')).then( (value)=> {
-      if (!value){
-        return this.handleUnauthorizedRequest(transition);
-      }
-    })
-  },
+  subRouteName: 'workflow'
 });