+++ CHANGELOG.md
@@ -4,6 +4,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 
 ## {[version]} - {release date}
 ### Added
+- Autosuggested at-mentions in discussion topics
 ### Changed
 ### Deprecated
 ### Removed